# api/swagger.json
{
	"openapi": "3.0.0",
	"paths": {
		"/chats": {
			"post": {
				"operationId": "ChatsController_createChat",
				"summary": "",
				"description": "метод создания чата",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateOpenChatDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean",
											"description": "является ли запрос успешным"
										},
										"data": {
											"$ref": "#/components/schemas/ChatEntity"
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"chats"
				]
			},
			"get": {
				"operationId": "ChatsController_getChats",
				"summary": "",
				"description": "метод получения чатов",
				"parameters": [
					{
						"name": "title",
						"required": false,
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "limit",
						"required": false,
						"in": "query",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "offset",
						"required": false,
						"in": "query",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean",
											"description": "является ли запрос успешным"
										},
										"data": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/ChatEntity"
											}
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"chats"
				]
			}
		},
		"/chats/{id}": {
			"get": {
				"operationId": "ChatsController_getChat",
				"summary": "",
				"description": "метод получения чата по id",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean",
											"description": "является ли запрос успешным"
										},
										"data": {
											"$ref": "#/components/schemas/ChatEntity"
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"chats"
				]
			}
		},
		"/messages": {
			"post": {
				"operationId": "MessagesController_createMessage",
				"summary": "",
				"description": "метод создания сообщения",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMessageDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean",
											"description": "является ли запрос успешным"
										},
										"data": {
											"$ref": "#/components/schemas/MessageEntity"
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"messages"
				]
			},
			"get": {
				"operationId": "MessagesController_getMessages",
				"summary": "",
				"description": "метод получения сообщений",
				"parameters": [
					{
						"name": "chatId",
						"required": false,
						"in": "query",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "limit",
						"required": false,
						"in": "query",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "offset",
						"required": false,
						"in": "query",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "parentMessageId",
						"required": true,
						"in": "query",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "search",
						"required": true,
						"in": "query",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean",
											"description": "является ли запрос успешным"
										},
										"data": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/MessageEntity"
											}
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"messages"
				]
			}
		}
	},
	"info": {
		"title": "passim",
		"description": "",
		"version": "0.0.1",
		"contact": {},
		"license": {
			"name": "UNLICENSED",
			"url": "https://github.com/ramzini22"
		}
	},
	"tags": [],
	"servers": [
		{
			"url": "http://localhost:6020",
			"description": "development"
		}
	],
	"components": {
		"schemas": {
			"ChatEntity": {
				"type": "object",
				"properties": {
					"id": {
						"type": "number"
					},
					"createdAt": {
						"format": "date-time",
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"countMessages": {
						"type": "number"
					},
					"createdUserId": {
						"type": "number"
					},
					"type": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"createdAt",
					"title",
					"countMessages",
					"createdUserId",
					"type"
				]
			},
			"CreateOpenChatDto": {
				"type": "object",
				"properties": {
					"title": {
						"type": "string"
					}
				},
				"required": [
					"title"
				]
			},
			"MessageEntity": {
				"type": "object",
				"properties": {
					"id": {
						"type": "number"
					},
					"createdAt": {
						"format": "date-time",
						"type": "string"
					},
					"encryptMessage": {
						"type": "string"
					},
					"chatId": {
						"type": "number"
					},
					"number": {
						"type": "number"
					},
					"message": {
						"type": "string"
					},
					"parentMessageId": {
						"type": "number"
					}
				},
				"required": [
					"id",
					"createdAt",
					"encryptMessage",
					"chatId",
					"number",
					"message",
					"parentMessageId"
				]
			},
			"CreateMessageDto": {
				"type": "object",
				"properties": {
					"encryptMessage": {
						"type": "string"
					},
					"chatId": {
						"type": "number"
					},
					"message": {
						"type": "string"
					},
					"parentMessageId": {
						"type": "number"
					}
				},
				"required": [
					"encryptMessage",
					"chatId",
					"message",
					"parentMessageId"
				]
			}
		}
	}
}